refactor(M): rename Simulation to Session in frontend for consistency with backend
