+++ frontend/src/app/[locale]/(standalone)/session/[id]/hooks/useMediaStream.ts
