+++ frontend/src/app/[locale]/(standalone)/session/[id]/hooks/useRemoteAudioRecorder.ts
