+++ frontend/src/app/[locale]/(standalone)/session/[id]/loading.tsx
