+++ frontend/src/app/[locale]/(standalone)/session/[id]/page.tsx
@@ -3,7 +3,7 @@ import type { Metadata } from 'next';
 import { MetadataProps } from '@/interfaces/props/MetadataProps';
 import { PagesProps } from '@/interfaces/props/PagesProps';
 import { getTranslations } from 'next-intl/server';
-import SimulationPageComponent from './components/SimulationPage';
+import SessionPageComponent from './components/SessionPage';
 
 export async function generateMetadata({ params }: MetadataProps): Promise<Metadata> {
   const { locale } = await params;
@@ -16,7 +16,7 @@ export default async function SimulationPage(props: PagesProps) {
 
   return (
     <div>
-      <SimulationPageComponent sessionId={id} />
+      <SessionPageComponent sessionId={id} />
       <div className="mt-4 mb-4 text-center">
         <p className="text-xs text-bw-40">{tCommon('aiDisclaimer')}</p>
       </div>
