--- frontend/src/interfaces/models/SessionLiveFeedback.ts
@@ -1,5 +1,0 @@
-export interface SessionLiveFeedback {
-  id: string;
-  heading: string;
-  feedbackText: string;
-}
